Group Api service
