+++ APILayer/Services/Implementations/APIService.cs
@@ -13,6 +13,7 @@ namespace APILayer.Services.Implementations
             _context = context;
         }
 
+        #region API
         public async Task<IEnumerable<API>> GetAllAPIsAsync()
         {
             return await _context.Set<API>().Include(a => a.Versions)
@@ -61,5 +62,96 @@ namespace APILayer.Services.Implementations
             await _context.SaveChangesAsync();
             return true;
         }
+        #endregion
+
+        #region APIDocumentation
+        public async Task<IEnumerable<APIDocumentation>> GetAllDocumentationsAsync()
+        {
+            return await _context.Set<APIDocumentation>().ToListAsync();
+        }
+
+        public async Task<APIDocumentation?> GetDocumentationByIdAsync(int id)
+        {
+            return await _context.Set<APIDocumentation>().FindAsync(id);
+        }
+
+        public async Task<APIDocumentation> CreateDocumentationAsync(APIDocumentation documentation)
+        {
+            await _context.Set<APIDocumentation>().AddAsync(documentation);
+            await _context.SaveChangesAsync();
+            return documentation;
+        }
+
+        public async Task<APIDocumentation?> UpdateDocumentationAsync(int id, APIDocumentation updatedDocumentation)
+        {
+            var existingDocumentation = await _context.Set<APIDocumentation>().FindAsync(id);
+            if (existingDocumentation == null)
+            {
+                return null;
+            }
+
+            _context.Entry(existingDocumentation).CurrentValues.SetValues(updatedDocumentation);
+            await _context.SaveChangesAsync();
+            return existingDocumentation;
+        }
+
+        public async Task<bool> DeleteDocumentationAsync(int id)
+        {
+            var documentation = await _context.Set<APIDocumentation>().FindAsync(id);
+            if (documentation == null)
+            {
+                return false;
+            }
+
+            _context.Set<APIDocumentation>().Remove(documentation);
+            await _context.SaveChangesAsync();
+            return true;
+        }
+        #endregion
+
+        #region APIVersion
+        public async Task<IEnumerable<APIVersion>> GetAllVersionsAsync()
+        {
+            return await _context.Set<APIVersion>().ToListAsync();
+        }
+
+        public async Task<APIVersion?> GetVersionByIdAsync(int id)
+        {
+            return await _context.Set<APIVersion>().FindAsync(id);
+        }
+
+        public async Task<APIVersion> CreateVersionAsync(APIVersion version)
+        {
+            await _context.Set<APIVersion>().AddAsync(version);
+            await _context.SaveChangesAsync();
+            return version;
+        }
+
+        public async Task<APIVersion?> UpdateVersionAsync(int id, APIVersion updatedVersion)
+        {
+            var existingVersion = await _context.Set<APIVersion>().FindAsync(id);
+            if (existingVersion == null)
+            {
+                return null;
+            }
+
+            _context.Entry(existingVersion).CurrentValues.SetValues(updatedVersion);
+            await _context.SaveChangesAsync();
+            return existingVersion;
+        }
+
+        public async Task<bool> DeleteVersionAsync(int id)
+        {
+            var version = await _context.Set<APIVersion>().FindAsync(id);
+            if (version == null)
+            {
+                return false;
+            }
+
+            _context.Set<APIVersion>().Remove(version);
+            await _context.SaveChangesAsync();
+            return true;
+        }
+        #endregion
     }
 }
